move some files to shared directory
