+++ server/src/main/java/edu/duke/ece651/risk_game/server/v1MapFactory.java
@@ -1,5 +1,5 @@
 package edu.duke.ece651.risk_game.server;
-
+import edu.duke.ece651.risk_game.shared.*;
 import java.util.List;
 
 public class v1MapFactory implements MapFactory{
