+++ server/src/main/java/edu/duke/ece651/risk_game/server/v1WorldMap.java
@@ -1,5 +1,5 @@
 package edu.duke.ece651.risk_game.server;
-
+import edu.duke.ece651.risk_game.shared.*;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
