+++ shared/src/main/java/edu/duke/ece651/risk_game/shared/CombatResolver.java
@@ -1,4 +1,4 @@
-package edu.duke.ece651.risk_game.server;
+package edu.duke.ece651.risk_game.shared;
 public interface CombatResolver {
     // return the number of units that survive, positive indicate attacker wins, negative indicate defender wins
     // the absolute value of return value indicate the number of units that survive
